feat: add student package and files
